Moved RandomColorTest to util package
